+++ src/Service/PUDOListStreaming.php
@@ -11,7 +11,7 @@ use Answear\DpdPlPickupServicesBundle\ValueObject\PUDO;
 use GuzzleHttp\Client;
 use GuzzleHttp\ClientInterface;
 
-class PUDOListStreaming
+class PUDOListStreaming extends AbstractPUDOList
 {
     private ClientInterface $client;
 
@@ -32,70 +32,8 @@ class PUDOListStreaming
 
     /**
      * @return iterable<PUDO>
-     *
-     * @throws ServiceException
      */
-    public function byAddress(string $zipCode, string $city, ?string $address = null): iterable
-    {
-        $params = [
-            'requestID' => \uniqid('', true),
-            'city' => $city,
-            'zipCode' => $zipCode,
-            'servicePudo_display' => 1,
-        ];
-        if (null !== $address) {
-            $params['address'] = $address;
-        }
-
-        return $this->request('list/byaddress', $params);
-    }
-
-    /**
-     * @return iterable<PUDO>
-     *
-     * @throws ServiceException
-     */
-    public function byCountry(string $countryCode): iterable
-    {
-        return $this->request('list/bycountry', ['countryCode' => $countryCode]);
-    }
-
-    /**
-     * @throws ServiceException
-     */
-    public function byId(string $id): ?PUDO
-    {
-        $generator = $this->request('details', ['pudoId' => $id]);
-
-        foreach ($generator as $pudo) {
-            return $pudo;
-        }
-
-        return null;
-    }
-
-    /**
-     * @return iterable<PUDO>
-     *
-     * @throws ServiceException
-     */
-    public function byLatLng(Coordinates $coordinates, int $distance): iterable
-    {
-        return $this->request(
-            'list/bylonglat',
-            [
-                'requestID' => \uniqid('', true),
-                'latitude' => $coordinates->latitude,
-                'longitude' => $coordinates->longitude,
-                'max_distance_search' => $distance,
-            ]
-        );
-    }
-
-    /**
-     * @return iterable<PUDO>
-     */
-    private function request(string $endpoint, array $params): iterable
+    protected function request(string $endpoint, array $params): iterable
     {
         $params['key'] = $this->configProvider->key;
